requirements.txt
bytewax==0.18
snowflake_connector